+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ### Unreleased
 * Add support for returning multiple messages by a list of message IDs
+* Add new field in `Draft` for adding files by file IDs
+* Add Webhook Notification models
 * Improved `Delta` support: added `Delta` model and two new methods; `since` and `longpoll`
+* Fix Virtual Calendar logic and serialization
 
 ### 6.1.0 / 2022-01-28
 * Add support for `Event` to ICS